Refactor: Move Session Config Into Dedicated File
- Move SessionConfig Into new file named session_config
- Move AccessMode and its constants into session_config file
- Move AuthToken type alias to session_config file
